aws: Rename `eni` package to `ipam`
Rename `pkg/aws/eni` to `pkg/aws/ipam` so the operator-side IPAM allocator
(the `ipamTypes.NodeOperations` implementation that runs in the Cilium
Operator) lives under a name that describes its role and is consistent
with other cloud provider packages
(`pkg/$cloud/{api,metadata,types,ipam}` layout).
