+++ pkg/aws/ipam/doc.go
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-// Package eni implements AWS ENI allocation logic
-package eni
+// Package ipam implements AWS ENI allocation logic
+package ipam
